requirements.txt
babel
backrefs
certifi
charset-normalizer
click
colorama
ghp-import
idna
importlib_metadata
Jinja2
Markdown
MarkupSafe
mergedeep
mkdocs
mkdocs-get-deps
mkdocs-material
mkdocs-material-extensions
packaging
paginate
pathspec
platformdirs
Pygments
pymdown-extensions
PyMuPDF  # Used for PDF text and image extraction
Pillow  # For image processing
pyyaml  # For updating mkdocs.yml
python-dateutil
PyYAML
pyyaml_env_tag
requests
six
urllib3
watchdog
zipp